haroon did small changes

+++ src/app/components/navbar/navbar.component.ts
@@ -7,6 +7,7 @@ import { DataService } from 'src/app/services/data.service';
 import { GetDataService } from 'src/app/services/getdata.service';
 import { DomSanitizer, SafeResourceUrl, SafeUrl} from '@angular/platform-browser';
 import { EmitterService } from '../../services/emitter.service';
+import { ToastrService } from 'ngx-toastr';
 
 
 @Component({
@@ -26,7 +27,9 @@ export class NavbarComponent implements OnChanges {
   imageurl: any;
   customer:any = [];
   productname;
-  constructor(public domSanitizer: DomSanitizer,public eventemitter : EmitterService,public route:Router,public cartservice:GetDataService, location: Location,  private element: ElementRef, private router: Router, private cookies: CookieService, private dataService: DataService) {
+  constructor(public domSanitizer: DomSanitizer,public eventemitter : EmitterService,public route:Router,public cartservice:GetDataService,
+     location: Location,  private element: ElementRef, private router: Router, private cookies: CookieService,
+      private dataService: DataService,private toast:ToastrService) {
     this.location = location;
     this.cartcount = this.cartservice.cartdata.count ;
    
@@ -61,15 +64,26 @@ export class NavbarComponent implements OnChanges {
   }
   logout()
   {
+    debugger
     let customer = JSON.parse(localStorage.getItem('customer')) ;
     let data = {"userno": customer.userno};
-    this.dataService.logout(data).subscribe((data)=>{
+    this.dataService.logout(data).subscribe((data:any)=>{
       debugger
+      if(data.errorstatusno == "1")
+      {
       this.cookies.deleteAll();
       this.cartservice.customer.customerdata = [];
       localStorage.removeItem('customer');
       this.cartservice.customer.isLogin = false;
       localStorage.removeItem('isLogin');
+      this.toast.success("Successfully Logout");
+      }
+      else
+{
+  this.toast.error(data.errortext);
+  
+
+}
     })
     
   // this.reload();
